Improved softphone error reporting.
Changes for VuMeter.

Make sendMessage() public instead of protected in VoiceBridgeConnection.

Added listeners to get voice bridge audio and rtp data.

Added a flag to AudioGroupSetup to not remove the group automatically
when the last member leaves the group.

Print a warning message if the voice bridge public address is unusable.




git-svn-id: https://svn.java.net/svn/jvoicebridge~svn/branches/jp@340 45c97d41-3a3b-0410-9611-e53bcf13ea59